Make QTT and VNN optional package extensions
