+++ docs/Project.toml
@@ -7,6 +7,8 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 TwoBody = "a92d7657-722c-45a6-9d18-9da4c8a753b6"
+TensorTrainNumerics = "304e70ee-b558-420e-aed5-de4d05b6ca2b"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 Antique = "0.15"
