+++ docs/make.jl
@@ -1,4 +1,8 @@
 using TwoBody
+import Lux
+import Optimisers
+import TensorTrainNumerics
+import Zygote
 using Documenter
 using DocumenterMermaid
